requirements.txt
altair==5.4.1
elevenlabs==1.12.1
groq==0.11.0
gTTS==2.5.3
langchain_core==0.3.15
langchain_google_genai==2.0.4
langchain_openai==0.2.6
pandas==2.2.3
pillow==10.4.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.39.0
streamlit_js_eval==0.1.7
